fix : user cash Integer->Long 변경

+++ api/src/main/java/io/eagle/domain/user/dto/response/UserInfoDto.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 public class UserInfoDto {
 
     private String username;
-    private Integer cash;
+    private Long cash;
     private Integer value;
     private String email;
     private ProviderType providerType;
